Continuing the Game - GameLogic and Input
Created the GameLogic class which will handle, as the name entails, most of the game's interactions. Right now, it creates 2 number-squares (2 or 4) at random non-occupied indexes when initialized; it adds a new square with every movement (although the movement  itself isn't yet coded); it keeps track of all the squares on the board; and it takes care of the restart, clearing the board and score before initializing itself again (which happens automatically if the board is full, but can be also be called as a command).

Also created a temporary player input to test out the game logic.