Continuing the Game - Squares Movement
Started coding the GameLogic's "Swipe" functions, which will deal with the  numbered squares movement on the board, as well as collisions. As of now, nothing happens when pieces collide yet, but the squares can already be moved in all directions without overlapping with each other (while updating the m_FreeSquares so the game restarts once the board is filled still).