Continuing the Game - Collisions and More
Improved the GameLogic's Swipe functions as to collapse squares together when they have the same value and collide. The only issue with these functions now is that sometimes, squares will take into account, when calculating their movement, pieces that have been collapsed right before on the same swipe - stopping 1 square short of where they should;

Made the squares collapsing accumulate into a score, and created a logic component for the score display, that now updates said display in real time;

Made the game only restart automatically when there's no possible movements that will free up space in the board (unlike before, where it would do it whenever the board was full and a swipe function was called, even if there were still possible moves left);

And lastly, made the NumberSquare class responsible for updating the square's graphics and text components when moving (instead of the GameLogic), which also led me to code a way to properly centre the square's text depending on how many chars it has.