Starting the AI - Non-Input Movement
With the base game done and bug-clean, I started developing the early stages of the AI.

To start off, I created the main class which will be responsible for outputting the movements to the GameLogic, with an adjustable interval between "swipes" (which can be dialled back to 0 later on for time-saving purposes during training). As of now, the class isn't really an AI yet, as it only plays randomly.

While preparing this, I realized it was important to make the game restart clearer and less sudden, as it used to happen automatically as soon as no more moves where available. Now, there's an adjustable pause with a "Game Over" title before the board and score reset.