Continuing the AI - Planning Training Structure
Implemented a WIP Train function to the PlayerAI which, when being called by the Update(), will not only execute the AI movement but also, as the name entails, gather information and optimize the AI playstyle. Right now, most of the functions the training depends on are still blank though.

Also changed some things in the GameLogic class for the AI's sake, including making the Swipe functions return the individual reward of each move, and making a getter function for the m_GameOver bool.