Continuing the AI - Expanding the PlayerAI Class
Continued expanding my PlayerAI class, although still in a very experimental and unrefined manner (and heavily inspired by other programmers projects of similar topics). As so, the current organization of this class is bound to change in the future and is the way it is now mostly as an aid for me to understand the concepts behind reinforcement learning.

Also created a function in the GameLogic class which returns the current state of the game - for now, as a 16 sized int vector with the values of all the squares in the board, as well as "0"s for the empty spots, sorted from left to right and top to bottom.