Continuing the AI - Implementing TorchLib
Implemented the TorchLib framework in the project as to facilitate the AI's neural network development. With the aid of this new library, I wrote a new class named LinearQNet which will not directly handle the agent's training, but will predict their moves when not choosing random outputs (for exploration's sake). Just like the last implementations in the PlayerAI class (as mentioned in the last commit's description), this class is still in an early state and its influences by other online projects I came across during my research are obvious - this will change in the future.

((As the inclusion of the library increases the size of the project immensely, I've added it to the .gitignore for now - I'll have to find a solution for this issue later))