Continuing the AI - Trying to Fix TorchLib and Others
First, tried to fix the TorchLib implementation as, once fully implemented, it started giving compile errors. This attempt did not pay off, but I intend on solving the issue as soon as possible.

Aside from that, now with a better understanding of the content I've been using as research and base for my PlayerAI's logic, I cleaned the class up and changed some of the overall organization to better suit my project's purpose (by deleting and adapting otherwise useless chunks of code, renaming variables and functions to more understandable titles, adding descriptive comments, etc);

Lastly, created a QLearningTrainer class, responsible for actually training the bot with the aid of the LinearQNet, and expanded this second class. Both of these are still in an untested and messy state, as I can't run TorchLib, and, therefore, are also still greatly influenced by my research material - as I haven't yet fully grasped some of the concept they entail.