Continuing the AI - Redoing Neural Networking
After a couple of days trying to find a way to properly implement TorchLib to no avail, I eventually removed the library from the project completely and programmed a substitute NeuralNetwork class from scratch. This new class was heavily influenced by another project I found during my research (as I point out in the header comments) and I intend on keeping it as such, but I did personalize it to my project's features - changing some of the structure, renaming some variables and pairing the rest of the unchanged code with personal comments explaining every step.
This new class made the LinearQNet one obsolete, and also led me to alter the QLearningTrainer class heavily,  as well as some chunks of the PlayerAI.