Continuing the AI - Training and Tweaking
Changed the GameLogic, the SquaredNumbers and the Factory class to allow games without visuals and minimal time constraints (for training purposes);
Added the total mean calculation of all scores to the game-over message printed every loop;
And did a couple of short training sessions (around 80 games each).
I noticed during these that the AI would start making only ilegal moves a bit after finishing its exploration phase, stagnating completely (thus stopping the sessions at 80). To try to fix this, I changed a couple of things in the NeuralNetwork class, as well as the PlayerAI, and altered the way the reward system and the environment, were fed to the network. The game state became a vector of the log2 of the values, instead of the full values (to uniformize the inputs), and the reward became the sum of the log2 of each collapsed cell with an extra -1 penalty for illegal moves, on top of the one for wasted moves (instead of just a +1 for each collapsed piece). This didn't fix the issue though, as the AI still stagnates around the 80 games milestone - so I capped the decrease of the random factor at 10% chance, and increased its initial value to 50%.